requirements.txt
fastapi==0.109.2
uvicorn[standard]
python-dotenv
pydantic==2.3.0
python-multipart==0.0.9
sqlalchemy==2.0.20
alembic== 1.12.0
psycopg2-binary
pytest
httpx==0.26.0
pytest-cov
email-validator
python-jose
boto3
bcrypt
passlib